requirements.txt
comet_ml
torch==1.8.0
torchvision
matplotlib
pillow
pandas
tqdm
numba
sklearn
pyyaml
albumentations